# openapi/v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Amortized Runtime",
    "description": "AI model customization runtime API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/jobs": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get Jobs",
        "operationId": "get_jobs_api_v1_jobs_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/JobStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/JobType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Job"
                  },
                  "title": "Response Get Jobs Api V1 Jobs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Create Job Universal",
        "operationId": "create_job_universal_api_v1_jobs_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/Job"
                    },
                    {
                      "$ref": "#/components/schemas/DryRunResponse"
                    }
                  ],
                  "title": "Response Create Job Universal Api V1 Jobs Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get Job Detail",
        "operationId": "get_job_detail_api_v1_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "jobs"
        ],
        "summary": "Cancel Job",
        "operationId": "cancel_job_api_v1_jobs__job_id__delete",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/jobs/validate": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Validate Job",
        "description": "Validate a job configuration without creating it.",
        "operationId": "validate_job_api_v1_jobs_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DryRunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/config/validate": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Validate Config Endpoint",
        "description": "Lightweight config validation \u2014 just type + config, no compute spec.",
        "operationId": "validate_config_endpoint_api_v1_config_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigValidateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigValidateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/job-types/{job_type}/schema": {
      "get": {
        "tags": [
          "job-types"
        ],
        "summary": "Get Job Type Schema",
        "operationId": "get_job_type_schema_api_v1_job_types__job_type__schema_get",
        "parameters": [
          {
            "name": "job_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Job Type Schema Api V1 Job Types  Job Type  Schema Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/job-types/{job_type}/examples": {
      "get": {
        "tags": [
          "job-types"
        ],
        "summary": "Get Job Type Examples",
        "description": "Return working config examples for a job type.",
        "operationId": "get_job_type_examples_api_v1_job_types__job_type__examples_get",
        "parameters": [
          {
            "name": "job_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Get Job Type Examples Api V1 Job Types  Job Type  Examples Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ComputeSpec": {
        "properties": {
          "backend": {
            "type": "string",
            "title": "Backend",
            "description": "Compute backend name (e.g. 'local', 'ssh')",
            "default": "local"
          },
          "gpus": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Gpus",
            "description": "Number of GPUs requested",
            "default": 0
          },
          "gpu_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gpu Type",
            "description": "GPU type (e.g. 'A100', 'H100')"
          }
        },
        "type": "object",
        "title": "ComputeSpec",
        "description": "Specifies which compute backend to use for a job."
      },
...
      "ConfigValidateRequest": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "description": "Job type (e.g. 'training', 'sdg')"
          },
          "config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Config",
            "description": "Job-type-specific configuration"
          }
        },
        "type": "object",
        "required": [
          "type",
          "config"
        ],
        "title": "ConfigValidateRequest",
        "description": "Lightweight config validation request (no compute spec)."
      },
      "ConfigValidateResponse": {
        "properties": {
          "valid": {
            "type": "boolean",
            "title": "Valid"
          },
          "errors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Errors"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "valid"
        ],
        "title": "ConfigValidateResponse",
        "description": "Response from config validation."
      },
...
      "DryRunResponse": {
        "properties": {
          "dry_run": {
            "type": "boolean",
            "title": "Dry Run",
            "default": true
          },
          "valid": {
            "type": "boolean",
            "title": "Valid"
          },
          "errors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Errors"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "compute": {
            "additionalProperties": true,
            "type": "object",
            "title": "Compute"
          },
          "config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Config"
          }
        },
        "type": "object",
        "required": [
          "valid",
          "type"
        ],
        "title": "DryRunResponse",
        "description": "Response from a dry-run / validation request."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Job": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "type": {
            "$ref": "#/components/schemas/JobType"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus",
            "default": "queued"
          },
          "config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Config"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "output_dir": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Dir"
          }
        },
        "type": "object",
        "required": [
          "type"
        ],
        "title": "Job",
        "description": "Job record returned by the API."
      },
      "JobRequest": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "description": "Job type (e.g. 'training', 'sdg')"
          },
          "config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Config",
            "description": "Job-type-specific configuration"
          },
          "compute": {
            "$ref": "#/components/schemas/ComputeSpec",
            "description": "Compute backend spec"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata",
            "description": "User-defined metadata"
          },
          "dry_run": {
            "type": "boolean",
            "title": "Dry Run",
            "description": "Validate and preview without creating the job",
            "default": true
          },
          "depends_on": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Depends On",
            "description": "Artifact references this job depends on (e.g. ['artifact:job-abc/model'])"
          }
        },
        "type": "object",
        "required": [
          "type",
          "config"
        ],
        "title": "JobRequest",
        "description": "Universal job submission request."
      },
      "JobStatus": {
        "type": "string",
        "enum": [
          "validating",
          "queued",
          "provisioning",
          "running",
          "succeeded",
          "failed",
          "cancelled",
          "queued",
          "succeeded"
        ],
        "title": "JobStatus",
        "description": "Status of a job."
      },
      "JobType": {
        "type": "string",
        "enum": [
          "training",
          "sdg",
          "inference",
          "eval"
        ],
        "title": "JobType",
        "description": "Type of job."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}